backend/requirements.txt
fastapi
sqlalchemy
alembic
databases
asyncpg
pydantic
sqlmodel
gspread
openai
python-dotenv